+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.34.0
 
+* Use JDK HTTP client in the Kubernetes client instead of the OkHttp client
+
 ## 0.33.0
 
 * Add support for Kafka 3.3.2